add another missing part of DB's patch to convert flags and related variables into longs: toggle_init_one() should take a long instead of an int for its flag parameter
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1845 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/global.c
@@ -215,9 +215,9 @@ void sc_init_one(shortcut **shortcutage, int ctrlval, const char *desc,
 }
 
 #ifndef NANO_SMALL
-/* Create one new toggle structure, at the end of the toggles
- * linked list. */
-void toggle_init_one(int val, const char *desc, int flag)
+/* Create one new toggle structure, at the end of the toggles linked
+ * list. */
+void toggle_init_one(int val, const char *desc, long flag)
 {
     toggle *u;
